Give up and start looking for kapt alternative
The problem is we are not in control of how the detector is created and
therefore we can't store information about the type of detector we are
creating, no related issue, no file rule whatsoever. I'm trying now with
kapt + kotlinpoet so that I can generate those classes on demand.